DRTII-1324 Start isolating crunch actors
