+++ server/src/main/scala/controllers/application/ApplicationInfoController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import buildinfo.BuildInfo
 import com.google.inject.Inject
 import drt.shared._
