+++ server/src/main/scala/controllers/application/ConfigController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import com.google.inject.Inject
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
