+++ server/src/main/scala/controllers/application/ContactDetailsController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import com.google.inject.Inject
 import drt.shared._
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
