+++ server/src/main/scala/controllers/application/DropInsController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import com.google.inject.Inject
 import drt.shared.{DropIn, DropInRegistration}
 import email.GovNotifyEmail
