+++ server/src/main/scala/controllers/application/EgateBanksController.scala
@@ -1,7 +1,6 @@
 package controllers.application
 
 import uk.gov.homeoffice.drt.actor.commands.Commands.GetState
-import actors.DrtSystemInterface
 import akka.pattern.ask
 import com.google.inject.Inject
 import drt.shared.CrunchApi.MillisSinceEpoch
