+++ server/src/main/scala/controllers/application/FeatureFlagsController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import com.google.inject.Inject
 import drt.shared.FeatureFlags
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
