+++ server/src/main/scala/controllers/application/WalkTimeController.scala
@@ -1,6 +1,5 @@
 package controllers.application
 
-import actors.DrtSystemInterface
 import com.google.inject.Inject
 import drt.shared.api.{WalkTime, WalkTimes}
 import play.api.mvc.{Action, AnyContent, ControllerComponents}
