+++ server/src/main/scala/services/ApiService.scala
@@ -1,7 +1,6 @@
 
 package services
 
-import actors.DrtSystemInterface
 import actors.PartitionedPortStateActor.GetStateForTerminalDateRange
 import actors.persistent.staffing.UpdateShifts
 import akka.actor._
